refactor: split tests into test/ directory
