+++ package.json
@@ -4,7 +4,7 @@
   "description": "Deskbookers API JavaScript SDK",
   "main": "dist/index.js",
   "scripts": {
-    "test": "ava --timeout=10s test.js",
+    "test": "ava --timeout=10s",
     "build": "babel src --out-dir dist",
     "prepublish": "npm run build"
   },
@@ -40,6 +40,10 @@
     "babel-preset-stage-2": "^6.18.0"
   },
   "ava": {
+    "require": [
+      "babel-register",
+      "babel-polyfill"
+    ],
     "babel": "inherit"
   }
 }
